[FIX] mail: public page fix
